+++ src/app/classes/interfaces.ts
@@ -1,7 +1,7 @@
 import { ChartData } from 'chart.js';
 import { Profile } from './profile';
 import { Graphs } from './constants';
-import { pages } from '../components/Help-Page/help-dialog/help-page-utils';
+import { pages } from '../help-dialog/help-dialog/help-page-utils';
 export interface IContributors {
   name: string;
   food_amount: number;
@@ -47,5 +47,5 @@ export interface IPageData {
 export interface IBillEntry {
   item: string;
   amount: number;
-  quantity?:number;
+  quantity?: number;
 }
