+++ src/app/help-dialog/about-page/about-page.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/palette.scss";
+@import "../../styles/palette.scss";
 .about-content {
   height: 100%;
   display: flex;
