+++ src/app/help-dialog/about-page/about-page.component.ts
