+++ src/app/help-dialog/help-dialog/help-dialog.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/palette.scss";
+@import "../../styles/palette.scss";
 $text-color: #8f9bb0;
 .help-wrapper {
   display: flex;
