+++ src/app/help-dialog/help-dialog/help-page-utils.ts
