+++ src/app/help-dialog/instruction-slides/details-panel-help/details-panel-help.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/palette.scss";
+@import "../../../styles/palette.scss";
 .slide-5 {
   .img-wrapper {
     display: flex;
