+++ src/app/help-dialog/instruction-slides/details-panel-help/details-panel-help.component.ts
