+++ src/app/help-dialog/instruction-slides/food-palette-help/food-palette-help.component.html
