+++ src/app/help-dialog/instruction-slides/food-palette-help/food-palette-help.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/palette.scss";
+@import "../../../styles/palette.scss";
 .slide-3 {
   .img-wrapper {
     text-align: center;
