+++ src/app/help-dialog/instruction-slides/getting-started/getting-started.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/palette.scss";
+@import "../../../styles/palette.scss";
 .slide-1 {
   .img-wrapper {
     text-align: center;
