+++ src/app/help-dialog/instruction-slides/getting-started/getting-started.component.ts
