+++ src/app/help-dialog/instruction-slides/palette-profiling-help/palette-profiling-help.component.scss
