+++ src/app/help-dialog/instruction-slides/profile-panel-help/profile-panel-help.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/palette.scss";
+@import "../../../styles/palette.scss";
 .slide-2 {
   p {
     line-height: 1.6;
