+++ src/app/help-dialog/instruction-slides/profile-panel-help/profile-panel-help.component.ts
