+++ src/app/help-dialog/instructions-page/instructions-page.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/palette.scss";
+@import "../../styles/palette.scss";
 .instructions-container {
   height: 100%;
   display: flex;
