+++ src/app/help-dialog/instructions-page/instructions-page.component.ts
