+++ src/app/help-dialog/keyboard-shortcuts/keyboard-shortcuts.component.html
