+++ src/app/help-dialog/release-notes/release-notes.component.scss
